Feature: Show links to equipment for the sport types.

+++ app/src/main/java/com/atrainingtracker/trainingtracker/database/EquipmentDbHelper.java
@@ -424,7 +424,7 @@ public class EquipmentDbHelper extends SQLiteOpenHelper {
     }
 
     @Nullable
-    public String getEquipmentNameFromId(int equipmentId) {
+    public String getEquipmentNameFromId(long equipmentId) {
         String equipmentName = null;
 
         SQLiteDatabase db = this.getReadableDatabase();
